Renamed fisher_vector.py to avoid circular import
