+++ skimage/feature/_fisher_vector.py
